refactor(kernel-core): move stripTypeScript to shared module, apply in runtime-quickjs too
Originally scoped to runtime-dynamic-worker but the same TS-strip step
applies to any JavaScript-only sandbox. QuickJS doesn't understand TS
either, so a typed annotation in user code would surface as a generic
QuickJsExecutionError instead of a clear syntax error. Lift
stripTypeScript into kernel/core (alongside recoverExecutionBody),
re-export from @executor-js/codemode-core, and call it from both
runtime-dynamic-worker and runtime-quickjs. runtime-deno-subprocess
still skips because Deno parses TS natively.

Tests move with the implementation (kernel/core/src/strip-types.test.ts).
Sucrase dep moves from runtime-dynamic-worker to kernel/core.
