+++ packages/kernel/core/src/index.ts
@@ -3,3 +3,4 @@ export * from "./validation";
 export * from "./json-schema";
 export * from "./effect-errors";
 export * from "./code-recovery";
+export * from "./strip-types";
