+++ packages/kernel/core/src/strip-types.test.ts
