+++ packages/kernel/runtime-dynamic-worker/package.json
@@ -14,8 +14,7 @@
   },
   "dependencies": {
     "@executor-js/codemode-core": "workspace:*",
-    "effect": "catalog:",
-    "sucrase": "^3.35.1"
+    "effect": "catalog:"
   },
   "devDependencies": {
     "@cloudflare/vitest-pool-workers": "^0.14.2",
